requirements.txt
flask 
flask-cors 
web3 
python-dotenv 
openai 
psutil 
groq 
requests 
